refactor: move legal documents to shared components/legals directory
- Move all privacy and agreement MDX files to components/legals/
- Update import statements to use shared @/components/legals/ path
- Remove empty local component directories
- Use plural directory name 'legals' for better naming convention

Files moved:
- en-privacy.mdx, zh-cn-privacy.mdx
- en-agreement.mdx, zh-cn-agreement.mdx

Now all legal document components are centralized and reusable.
