+++ src/cook-web/src/app/agreement/page.tsx
@@ -1,6 +1,6 @@
 'use client';
-import ZH_CN_Agreement from './components/zh-cn-agreement.mdx';
-import EN_Agreement from './components/en-agreement.mdx';
+import ZH_CN_Agreement from '@/components/legals/zh-cn-agreement.mdx';
+import EN_Agreement from '@/components/legals/en-agreement.mdx';
 
 import i18n from '@/i18n';
