+++ src/cook-web/src/components/legals/en-agreement.mdx
