+++ src/cook-web/src/components/legals/en-privacy.mdx
