+++ src/cook-web/src/components/legals/zh-cn-agreement.mdx
