+++ src/cook-web/src/components/legals/zh-cn-privacy.mdx
